DataMagic class -> module refactor into a few files for readability
